store original Items array so we don't mark modded items as vanilla on re-host
